app/requirements.txt
accelerate==1.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
decorator==5.1.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.1
platformdirs==4.3.6
pooch==1.8.2
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.41.2
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.5.1+cpu
tqdm==4.66.6
transformers==4.46.1
triton==3.1.0
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
